Move metadata classes into separate files.
